% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_information.R,
%   R/get_pipeline_report.R
\name{get_account_information}
\alias{get_account_information}
\title{get_account_information}
\usage{
get_account_information(user_code, api_token, pipelineid)

get_account_information(user_code, api_token, pipelineid)
}
\arguments{
\item{user_code}{The user code to identify your account}

\item{api_token}{The api token to connect to your account}

\item{user_code}{The user code to identify your account}

\item{api_token}{The api token to connect to your account}
}
\description{
Return your user account information from Less annoying CRM

Return your pipeline report information from Less annoying CRM.
For this you will need to know the pipelineId, StatusId, and CustomFieldId
You can get this PipelineId's at https://www.lessannoyingcrm.com/app/Settings/Api
}
