% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contact_information.R
\name{get_contact_information}
\alias{get_contact_information}
\title{get_contact_information}
\usage{
get_contact_information(user_code, api_token, search_term = "")
}
\arguments{
\item{user_code}{The user code to identify your account}

\item{api_token}{The api token to connect to your account}

\item{contact_id}{The contact name or other term to make an specific call
to the API.}
}
\description{
Return the contact information.
}
