% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pipeline_report.R
\name{get_pipeline_report}
\alias{get_pipeline_report}
\title{get_pipeline_report}
\usage{
get_pipeline_report(user_code, api_token, pipelineid)
}
\arguments{
\item{user_code}{The user code to identify your account}

\item{api_token}{The api token to connect to your account}

\item{pipelineid}{The id of the pipeline you want to get the report from}
}
\description{
Return your pipeline report information from Less annoying CRM.
}
\details{
For this you will need to know the pipelineId, StatusId, and CustomFieldId
You can get this PipelineId's at https://www.lessannoyingcrm.com/app/Settings/Api'
}
\examples{
\dontrun{
get_pipeline_report(user_code = "6A6E88",
                    api_token = "96066",
                    pipelineid = "57102821")
}

}
