% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contact_information.R
\name{get_test_contact_information}
\alias{get_test_contact_information}
\title{get_test_contact_information}
\usage{
get_test_contact_information(test_data)
}
\arguments{
\item{user_code}{}
}
\description{
Return the contact information to verify through test.
}
