% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacrmr-package.R
\docType{package}
\name{lacrmr-package}
\alias{lacrmr}
\alias{lacrmr-package}
\title{lacrmr: Connect to the 'Less Annoying CRM' API}
\description{
Connect to the 'Less Annoying CRM' API with ease to get your crm data in a clean and tidy format. 'Less Annoying CRM' is a simple CRM built for small businesses, more information is available on their website \url{https://www.lessannoyingcrm.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ixpantia.github.io/lacrmr/}
  \item Report bugs at \url{https://github.com/ixpantia/lacrmr/issues}
}

}
\author{
\strong{Maintainer}: Frans van Dunné \email{frans@ixpantia.com} (\href{https://orcid.org/0000-0002-7853-2811}{ORCID})

Authors:
\itemize{
  \item Ronny Hernández Mora \email{ronny@ixpantia.com} (\href{https://orcid.org/0000-0001-6225-7096}{ORCID})
}

Other contributors:
\itemize{
  \item ixpantia, family = SRL \email{hola@ixpantia.com} [copyright holder]
}

}
\keyword{internal}
