% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_information.R
\name{get_account_information}
\alias{get_account_information}
\title{get_account_information}
\usage{
get_account_information(user_code, api_token)
}
\arguments{
\item{user_code}{The user code to identify your account}

\item{api_token}{The api token to connect to your account}
}
\description{
Return your user account information from Less annoying CRM
}
\details{
For using this function you will need to get your credentials from
your Less Annoying CRM. Make sure to have your user code and your api_token.
}
\examples{
\dontrun{
get_account_information(user_code = "6A6E88",
                       api_token = "FBHV7C")
}

}
